wip: add notmuch-progressive-search custom
This change updates the process-filter function to send a SIGSTOP
signal to the notmuch-search process when the window is close to
the end of the buffer. Then a scroll-functions is used to send a
SIGCONT signal when the window reaches the end of the buffer.

This change also adds a notmuch-flush-buffer function to read the
whole search process outputs.

This change also ignore the SIGCHLD handler for the notmuch-search
process to properly reap gpg child process.